+++ app/Models/PatronType.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+
+class PatronType extends Model
+{
+    use HasFactory;
+    use SoftDeletes;
+
+    protected $guarded = [];
+
+    public function patrons() {
+        return $this->hasMany(Patron::class, 'patron_type');
+    }
+
+    public function addedBy() {
+        return $this->belongsTo(User::class,'added_by');
+    }
+}