+++ database/migrations/2026_04_29_040015_create_librarians_table.php
@@ -1,0 +1,38 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('librarians', function (Blueprint $table) {
+            $table->id();
+            $table->unsignedBigInteger('user_id');
+            $table->foreign('user_id')->references('id')->on('users')->onDelete('restrict')->onUpdate('cascade');
+            $table->string('librarian_code');
+            $table->string('last_name');
+            $table->string('first_name');
+            $table->string('middle_name')->nullable(true);
+            $table->string('suffix')->nullable(true);
+            $table->enum('gender', ['Male','Female']);
+            $table->string('profile_picture')->nullable(true);
+            $table->string('contact_number');
+            $table->softDeletes();
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('librarians');
+    }
+};