+++ changelog.md
@@ -1,4 +1,9 @@
 # Changelogs
 
+**v0.0.1 - April 29, 2026**
+- Instaled Laravel Breeze
+- Add and Update Seeders and Migrations
+- Initiate Auth Controllers
+
 **v0.0.0 - April 28, 2026**
 - Initialize Repository